+++ changelog.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## Upcoming
+## Beta 3.0.2
 * Now requires InsaneLib 1.2.1
 * `difficulty_modifier` is now specified in Range objects (min max) and in chance objects
-* Added `pos_modifier` in Range objects (min max) and in chance objects
+* Chance is now a modifiable value object. Now uses `value` instead of `amount`
+* Added `pos_modifier` in Range objects (min max) and in modifiable values objects (chance and weight). Item no longer has a `world_difficulty` modifier (as it's in the Weight object)
 * Added Phantom Size property
 * Moved `dimensions` and `biomes` whitelist to a `world_whitelist` object
 * Fixed Slot not checking for world whitelist correctly